requirements.txt
bcrypt==2.0.0
cffi==1.3.1
decorator==4.0.4
Flask==0.10.1
Flask-Bcrypt==0.7.1
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
pbr==1.8.1
pycparser==2.14
six==1.10.0
SQLAlchemy==1.0.9
sqlalchemy-migrate==0.10.0
sqlparse==0.1.18
Tempita==0.5.2
Werkzeug==0.11.2
WTForms==2.0.2
